requirements.txt
streamlit
torch
torchvision
pillow
streamlit-webrtc
av
